add more demos to example app
