+++ example-app/src/LogoDemo.js
@@ -1,0 +1,17 @@
+import React, { Component } from 'react'
+
+import Paper from 'ui/Paper';
+
+import Logo from 'ui/Logo/Logo';
+
+export default class ProgressDemo extends Component {
+    render() {
+        return (
+            <Paper elevation={6} padded>
+                <h3>Logotype components</h3>
+                <hr/>
+                <Logo icon />
+            </Paper>
+        )
+    }
+}
